Ah. One should not assume that uint is a 64-bit integer.

+++ bitmapcontainer.go
@@ -124,7 +124,7 @@ func (bc *bitmapContainer) add(i uint16) container {
 	x := int(i)
 	previous := bc.bitmap[x/64]
 	bc.bitmap[x/64] |= (1 << (uint(x) % 64))
-	bc.cardinality += int(uint(previous^bc.bitmap[x/64]) >> (uint(x) % 64))
+	bc.cardinality += int(uint64(previous^bc.bitmap[x/64]) >> (uint(x) % 64))
 	return bc
 }
 
@@ -307,7 +307,7 @@ func (bc *bitmapContainer) orArray(value2 *arrayContainer) container {
 	answer := bc.clone().(*bitmapContainer)
 	for k := 0; k < value2.getCardinality(); k++ {
 		i := uint(value2.content[k]) >> 6
-		answer.cardinality += int(uint(^answer.bitmap[i]&(1<<(value2.content[k]%64))) >> (value2.content[k] % 64))
+		answer.cardinality += int(uint64(^answer.bitmap[i]&(1<<(value2.content[k]%64))) >> (value2.content[k] % 64))
 		answer.bitmap[i] = answer.bitmap[i] | (uint64(1) << (value2.content[k] % 64))
 	}
 	return answer
@@ -326,11 +326,12 @@ func (bc *bitmapContainer) computeCardinality() {
 	bc.cardinality = int(popcntSlice(bc.bitmap))
 }
 
+
 func (bc *bitmapContainer) iorArray(value2 *arrayContainer) container {
 	answer := bc
 	for k := 0; k < value2.getCardinality(); k++ {
 		i := uint(value2.content[k]) >> 6
-		answer.cardinality += int(uint(^answer.bitmap[i]&(1<<(value2.content[k]%64))) >> (value2.content[k] % 64))
+		answer.cardinality += int(uint64(^answer.bitmap[i]&(1<<(value2.content[k]%64))) >> (value2.content[k] % 64))
 		answer.bitmap[i] = answer.bitmap[i] | (uint64(1) << (value2.content[k] % 64))
 	}
 	return answer
@@ -377,7 +378,7 @@ func (bc *bitmapContainer) xorArray(value2 *arrayContainer) container {
 	answer := bc.clone().(*bitmapContainer)
 	for k := 0; k < value2.getCardinality(); k++ {
 		index := uint(value2.content[k]) >> 6
-		answer.cardinality += 1 - 2*int(uint(answer.bitmap[index]&(1<<(value2.content[k]%64)))>>(value2.content[k]%64))
+		answer.cardinality += 1 - 2*int(uint64(answer.bitmap[index]&(1<<(value2.content[k]%64)))>>(value2.content[k]%64))
 		answer.bitmap[index] = answer.bitmap[index] ^ (uint64(1) << (value2.content[k] % 64))
 	}
 	if answer.cardinality <= arrayDefaultMaxSize {
@@ -513,7 +514,7 @@ func (bc *bitmapContainer) andNotArray(value2 *arrayContainer) container {
 	for k := 0; k < value2.getCardinality(); k++ {
 		i := uint(value2.content[k]) >> 6
 		answer.bitmap[i] = answer.bitmap[i] &^ (uint64(1) << (value2.content[k] % 64))
-		answer.cardinality -= int(uint(answer.bitmap[i]^bc.bitmap[i]) >> (value2.content[k] % 64))
+		answer.cardinality -= int(uint64(answer.bitmap[i]^bc.bitmap[i]) >> (value2.content[k] % 64))
 	}
 	if answer.cardinality <= arrayDefaultMaxSize {
 		return answer.toArrayContainer()
